requirements.txt
PyCryptodome
numpy
pillow
trio
pyfuse3